Fixed slider of 64bit control (CSIGENERAL-2689, OCT-3415)
The class QSlider can only handle 32bit value, but in the 64bit control element the 64bit value was passed without conversion to the slider. This caused an overflow of the internal slider value and the maximum value of the control element was not correct.
This is fixed by a new conversion, which converts the 64bit values with a fixed conversion factor to the 32bit slider values. For all 64bit exposure control, the slider now uses a logarithmic representation.