Fixed CSIGENERAL-2785: corrupt image for pixel format RGB3
The viewer was showing a corrupt images on zcu106 for some resolutions when the pixel format RGB3 or BGR3 were used. The images were not displayed correctly, because the viewer was not taking the bytes per line value into account. The xilinx drivers requires a certain byte alignment for each line.