Fixed CSIGENERAL-2829
When setting the cropping on the camera subdev, the image size was not correctly passed to the other subdevs. The original crop width and height were used and not the values returned by the driver. If a larger width or height value than the maximum of the camera subdev was entered, the wrong value was passed to the other subdevs without adjustments. This led to stream start errors on the AMD Xilinx boards.